skills/wechat-draft-publisher/requirements.txt
markdown>=3.5
beautifulsoup4>=4.12
requests>=2.31
pyyaml>=6.0
